+++ packages/client-wallets/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [typescript v1.0.0-alpha.4] packages/client-wallets
 ### Releases
 - [NPM v1.0.0-alpha.4] https://www.npmjs.com/package/@formance/sdk-wallets/v/1.0.0-alpha.4 - packages/client-wallets
+
+## 2024-07-17 00:11:38
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.337.1 (2.373.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.5] packages/client-wallets
+### Releases
+- [NPM v1.0.0-alpha.5] https://www.npmjs.com/package/@formance/sdk-wallets/v/1.0.0-alpha.5 - packages/client-wallets